I've created two fixed versions of the calculator.
At first when i tried to fix the issues i encountered with the %decimal% function, so that's what i toyed around with to make the calculator better. Then i realized that you advised to use the %round% function, so I just saved the file with the decimal function and then recreated the calculator so instead of working with decimals it just round the answer to 2. 

Of course beside fixing that, i also managed to clean the problem regarding dividing by zero crash and the bug on the next calculation.

If you ask me, I personally feel that the decimal calculator is faster than the one that uses the %round% function. May be is just my pc. 

Any way, the important is that I managed to learn more about two different functions along with all the operators and objects related to them in this project